raka: memindahkan fragment kedalam satu direktori bernama fragment
