+++ program/coralConnect/app/src/main/java/com/uti/coralconnect/fragment/NewstContentFragment.kt
@@ -1,0 +1,55 @@
+package com.uti.coralconnect.fragment
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.uti.coralconnect.ARG_PARAM1
+import com.uti.coralconnect.ARG_PARAM2
+import com.uti.coralconnect.R
+
+/**
+ * A simple [Fragment] subclass.
+ * Use the [NewstContentFragment.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+class NewstContentFragment : Fragment() {
+    private var param1: String? = null
+    private var param2: String? = null
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        arguments?.let {
+            param1 = it.getString(ARG_PARAM1)
+            param2 = it.getString(ARG_PARAM2)
+        }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_newst_content, container, false)
+    }
+
+    companion object {
+        /**
+         * Use this factory method to create a new instance of
+         * this fragment using the provided parameters.
+         *
+         * @param param1 Parameter 1.
+         * @param param2 Parameter 2.
+         * @return A new instance of fragment NewstContentFragment.
+         */
+        @JvmStatic
+        fun newInstance(param1: String, param2: String) =
+            NewstContentFragment().apply {
+                arguments = Bundle().apply {
+                    putString(ARG_PARAM1, param1)
+                    putString(ARG_PARAM2, param2)
+                }
+            }
+    }
+}
