+++ program/coralConnect/app/src/main/res/layout/fragment_camera.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CameraFragment">
+    tools:context=".fragment.CameraFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
